+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- `exp-expert-eta`: (new) Displays ETA for next expert to appear
+- `exp-expert-eta`: (new) Displays ETA for the next expert to appear
 - `show-space-remaining`: (new) Shows the remaining weight and volume capacity of the selected store in INV and SHPI
 - `wf-workforce-filters`: (new) Adds filters to hide zero workforce types and consumables